+++ src/main/java/com/dnd/MusicLog/music/controller/SpotifyMusicController.java
@@ -3,7 +3,7 @@ package com.dnd.MusicLog.music.controller;
 import com.dnd.MusicLog.global.common.BaseController;
 import com.dnd.MusicLog.global.jwt.util.JwtTokenProvider;
 import com.dnd.MusicLog.music.dto.SpotifyTrackResponseDto;
-import com.dnd.MusicLog.music.service.MusicService;
+import com.dnd.MusicLog.music.service.SpotifyMusicService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
@@ -17,7 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class SpotifyMusicController extends BaseController {
 
     private final JwtTokenProvider jwtTokenProvider;
-    private final MusicService musicService;
+    private final SpotifyMusicService spotifyMusicService;
 
     @GetMapping("")
     public SpotifyTrackResponseDto searchSpotifyMusic(
@@ -25,7 +25,7 @@ public class SpotifyMusicController extends BaseController {
         @RequestParam("query") String query,
         @RequestParam("offset") int offset) {
         jwtTokenProvider.extractAccessTokenSubject(token);
-        
-        return musicService.searchSpotifyMusic(query, offset);
+
+        return spotifyMusicService.searchSpotifyMusic(query, offset);
     }
 }
